+++ plugins/__init__.py
@@ -3,3 +3,5 @@ plugins
 =======
 """
 from . import modules
+
+__all__ = ["modules"]