+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.1.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Installs missing stubs automatically for `mypy==0.910`
+
+### Security
+- Upgrades package requirements
 
 [3.1.0](https://github.com/jshwi/pyaud/releases/tag/v3.1.0) - 2021-07-27
 ------------------------------------------------------------------------