+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.1.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Adds `__all__` to `plugins`
+
 ### Fixed
 - Installs missing stubs automatically for `mypy==0.910`